✨ [feature] 반응 남기기 기능 구현 및 나의 전시 뒷배경 수정
