+++ Projects/LastDance/LastDance/Sources/Domain/Mapper/ReactionMapper.swift
@@ -12,12 +12,27 @@ enum ReactionMapper {
         //            ReactionTagInfo(name: $0.name, colorHex: $0.color_hex ?? "#8F8F8F")
         //        }  // Fallback to gray
 
+        // 작가 이모지 추출 (첫 번째 이모지만 사용)
+        let artistEmoji = dto.artist_emojis?.first?.emoji_type
+
+        // 작가 메시지 매핑
+        let artistMessages = dto.artist_messages?.map {
+            ArtistMessageInfo(
+                id: $0.id,
+                artistName: $0.artist_name,
+                message: $0.message,
+                createdAt: $0.created_at
+            )
+        }
+
         return Reaction(
             id: String(dto.id),
             artworkId: dto.artwork_id,
             visitorId: dto.visitor_id,
             //            tags: tags,
             comment: dto.comment,
+            artistEmoji: artistEmoji,
+            artistMessages: artistMessages,
             createdAt: dto.created_at
         )
     }
