+++ Projects/LastDance/LastDance/Sources/Domain/Model/Reaction/Reaction.swift
@@ -13,6 +13,13 @@ struct ReactionTagInfo: Codable, Hashable {
     let colorHex: String
 }
 
+struct ArtistMessageInfo: Codable, Hashable {
+    let id: Int
+    let artistName: String
+    let message: String
+    let createdAt: String
+}
+
 @Model
 final class Reaction {
     var id: String
@@ -20,17 +27,22 @@ final class Reaction {
     var visitorId: Int
     var tags: [ReactionTagInfo]?
     var comment: String?
+    var artistEmoji: String?
+    var artistMessages: [ArtistMessageInfo]?
     var createdAt: String?
 
     init(
         id: String, artworkId: Int, visitorId: Int, tags: [ReactionTagInfo]? = nil,
-        comment: String? = nil, createdAt: String? = nil
+        comment: String? = nil, artistEmoji: String? = nil, artistMessages: [ArtistMessageInfo]? = nil,
+        createdAt: String? = nil
     ) {
         self.id = id
         self.artworkId = artworkId
         self.visitorId = visitorId
         self.tags = tags
         self.comment = comment
+        self.artistEmoji = artistEmoji
+        self.artistMessages = artistMessages
         self.createdAt = createdAt
     }
 }
