+++ Projects/LastDance/LastDance/Sources/Presentation/Reaction/Components/ArtworkReactionTab.swift
@@ -1,0 +1,11 @@
+//
+//  ArtworkReactionTab.swift
+//  LastDance
+//
+//  Created by donghee on 11/21/25.
+//
+
+enum ArtworkReactionTab {
+    case artwork
+    case message
+}
