+++ Projects/LastDance/LastDance/Sources/Presentation/Reaction/Views/ArtworkReactionView.swift
@@ -8,13 +8,6 @@
 import SwiftData
 import SwiftUI
 
-// MARK: ArtworkReactionTab
-
-private enum ArtworkReactionTab {
-    case artwork
-    case message
-}
-
 // MARK: ArtworkReactionView
 
 struct ArtworkReactionView: View {
@@ -72,20 +65,23 @@ struct ResponseContentView: View {
     @State private var selectedReactionForMessage: ReactionData?
 
     var body: some View {
-        ZStack {
-            VStack(spacing: 0) {
-                ZStack(alignment: .bottomLeading) {
-                    ArtworkBackgroundView(artwork: artwork)
-                        .frame(height: 393)
-                        .clipped()
-
+        ZStack(alignment: .top) {
+            // 배경 이미지
+            ArtworkBackgroundView(artwork: artwork)
+                .frame(height: 393)
+                .clipped()
+                .overlay(alignment: .bottomLeading) {
                     // 탭 바를 이미지 위에 배치
                     ResponseTabBar(selectedTab: $selectedTab)
                         .padding(.bottom, 24)
                 }
                 .ignoresSafeArea(edges: .top)
 
-                // 콘텐츠만 스크롤 (상단 블러 포함)
+            // 콘텐츠만 스크롤 (상단 블러 포함)
+            VStack(spacing: 0) {
+                Spacer()
+                    .frame(height: 393)
+
                 ZStack(alignment: .top) {
                     ScrollView {
                         if selectedTab == .artwork {
@@ -114,7 +110,7 @@ struct ResponseContentView: View {
                         startPoint: .top,
                         endPoint: .bottom
                     )
-                    .frame(height: 40)
+                    .frame(height: 20)
                     .allowsHitTesting(false)
                 }
             }
@@ -146,395 +142,20 @@ struct ResponseContentView: View {
                     }
                     .zIndex(20)
 
-                MessagePopupView(
-                    showMessagePopup: $showMessagePopup,
-                    viewModel: viewModel,
-                    reaction: selectedReactionForMessage
-                )
-                .transition(.opacity.combined(with: .scale))
+                GeometryReader { geometry in
+                    MessagePopupView(
+                        showMessagePopup: $showMessagePopup,
+                        viewModel: viewModel,
+                        reaction: selectedReactionForMessage
+                    )
+                    .position(
+                        x: geometry.size.width / 2,
+                        y: geometry.size.height / 2
+                    )
+                    .transition(.opacity.combined(with: .scale))
+                }
                 .zIndex(21)
             }
         }
-    }
-}
-
-// MARK: - ResponseTabBar
-
-struct ResponseTabBar: View {
-    @Binding fileprivate var selectedTab: ArtworkReactionTab
-
-    var body: some View {
-        HStack(spacing: 18) {
-            Button(action: {
-                selectedTab = .artwork
-            }) {
-                Text("작품")
-                    .font(LDFont.heading03)
-                    .foregroundColor(selectedTab == .artwork ? LDColor.color1 : LDColor.color2)
-            }
-            .buttonStyle(PlainButtonStyle())
-
-            Button(action: {
-                selectedTab = .message
-            }) {
-                Text("메시지")
-                    .font(LDFont.heading03)
-                    .foregroundColor(selectedTab == .message ? LDColor.color1 : LDColor.color2)
-            }
-            .buttonStyle(PlainButtonStyle())
-
-            Spacer()
-        }
-        .padding(.horizontal, 20)
-    }
-}
-
-// MARK: - ArtworkInfoSection
-
-struct ArtworkInfoSection: View {
-    let artwork: Artwork?
-    @ObservedObject var viewModel: ArtworkReactionViewModel
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            // 반응 수
-            Text("반응 수")
-                .font(LDFont.heading04)
-                .foregroundColor(LDColor.color1)
-                .padding(.top, 36)
-
-            Text("\(viewModel.reactions.count)")
-                .font(LDFont.heading01)
-                .foregroundColor(LDColor.color1)
-                .padding(.top, 8)
-
-            // 작품 제목
-            Text("작품 제목")
-                .font(LDFont.heading04)
-                .foregroundColor(LDColor.color1)
-                .padding(.top, 48)
-
-            Text(artwork?.title ?? "")
-                .font(LDFont.heading03)
-                .foregroundColor(LDColor.color1)
-                .padding(.top, 8)
-
-            // 작품 설명
-            Text("작품 설명")
-                .font(LDFont.heading04)
-                .foregroundColor(LDColor.color1)
-                .padding(.top, 36)
-
-            if let description = artwork?.descriptionText, !description.isEmpty {
-                Text(description)
-                    .font(LDFont.medium03)
-                    .foregroundColor(LDColor.color2)
-                    .lineSpacing(4)
-                    .padding(.top, 8)
-            } else {
-                Text("작품 설명이 없습니다.")
-                    .font(LDFont.medium03)
-                    .foregroundColor(LDColor.color2)
-                    .padding(.top, 8)
-            }
-
-            Spacer(minLength: 100)
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .padding(.horizontal, 20)
-    }
-}
-
-// MARK: - MessageListView
-
-struct MessageListView: View {
-    @ObservedObject var viewModel: ArtworkReactionViewModel
-    @Binding var showEmojiPopup: Bool
-    @Binding var emojiPopupPosition: CGRect
-    @Binding var showMessagePopup: Bool
-    @Binding var selectedReactionForMessage: ReactionData?
-
-    var body: some View {
-        VStack(spacing: 0) {
-            ForEach(0..<viewModel.reactions.count, id: \.self) { index in
-                MessageItemView(
-                    reaction: viewModel.reactions[index],
-                    index: index,
-                    viewModel: viewModel,
-                    isLast: index == viewModel.reactions.count - 1,
-                    showEmojiPopup: $showEmojiPopup,
-                    emojiPopupPosition: $emojiPopupPosition,
-                    showMessagePopup: $showMessagePopup,
-                    selectedReactionForMessage: $selectedReactionForMessage
-                )
-            }
-        }
-        .padding(.top, 20)
-        .padding(.bottom, 100)
-    }
-}
-
-// MARK: - MessageItemView
-
-struct MessageItemView: View {
-    let reaction: ReactionData
-    let index: Int
-    @ObservedObject var viewModel: ArtworkReactionViewModel
-    let isLast: Bool
-    @Binding var showEmojiPopup: Bool
-    @Binding var emojiPopupPosition: CGRect
-    @Binding var showMessagePopup: Bool
-    @Binding var selectedReactionForMessage: ReactionData?
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            VStack(alignment: .leading, spacing: 0) {
-                Text(Date.formatShortDate(from: reaction.createdAt))
-                    .font(LDFont.medium05)
-                    .foregroundColor(LDColor.color3)
-
-                Spacer().frame(height: 8)
-
-                // 댓글 텍스트
-                if !reaction.comment.isEmpty {
-                    VStack(alignment: .leading, spacing: 12) {
-                        Text(viewModel.displayText(for: reaction))
-                            .font(LDFont.medium03)
-                            .foregroundColor(LDColor.color1)
-                            .lineLimit(viewModel.expandedReactions.contains(reaction.id) ? nil : 3)
-                            .lineSpacing(6)
-
-                        // TODO: 글자수 말고 글자 높이, 너비로 접기-더보기 수정
-                        if reaction.comment.count > 80 {
-                            Button(action: {
-                                viewModel.handleExpandToggle(for: reaction)
-                            }) {
-                                Text(
-                                    viewModel.expandedReactions.contains(reaction.id) ? "접기" : "더보기"
-                                )
-                                .font(LDFont.medium05)
-                                .foregroundColor(LDColor.color3)
-                            }
-                        }
-                    }
-                    .padding(.bottom, 12)
-                }
-
-                HStack(spacing: 18) {
-                    GeometryReader { proxy in
-                        Button(action: {
-                            // 이미지 리스트에서 클릭된 이모지가 없다면 토글 가능
-                            if viewModel.getSelectedEmoji(for: reaction.id) == nil {
-                                if viewModel.selectedReactionId == reaction.id && showEmojiPopup {
-                                    // 이미 팝업이 열려있으면 닫기
-                                    showEmojiPopup = false
-                                    viewModel.selectedReactionId = nil
-                                } else {
-                                    // 팝업 열기
-                                    viewModel.selectedReactionId = reaction.id
-                                    emojiPopupPosition = proxy.frame(in: .global)
-                                    showEmojiPopup = true
-                                }
-                            }
-                        }) {
-                            // 선택된 이미지가 존재한다면 그걸로 표시
-                            if let selectedEmoji = viewModel.getSelectedEmoji(for: reaction.id) {
-                                Image(selectedEmoji)
-                                    .resizable()
-                                    .scaledToFill()
-                                    .frame(width: 31, height: 29)
-
-                            } else {
-                                Image(
-                                    viewModel.selectedReactionId == reaction.id && showEmojiPopup
-                                        ? "defaultImageFill" : "defaultImage"
-                                )
-                                .resizable()
-                                .scaledToFill()
-                                .frame(width: 26, height: 27)
-                            }
-                        }
-                        .disabled(viewModel.getSelectedEmoji(for: reaction.id) != nil)
-                    }
-                    .frame(width: 26, height: 27)
-
-                    Button(action: {
-                        selectedReactionForMessage = reaction
-                        showMessagePopup = true
-                    }) {
-                        Image(reaction.artistMessages.isEmpty ? "bubbleOff" : "bubbleOn")
-                            .resizable()
-                            .scaledToFill()
-                            .frame(width: 26, height: 25)
-                    }
-                    .buttonStyle(PlainButtonStyle())
-
-                    Spacer()
-                }
-
-                // 작가가 남긴 답글들
-                if !reaction.artistMessages.isEmpty {
-                    VStack(alignment: .leading, spacing: 8) {
-                        ForEach(reaction.artistMessages) { artistMessage in
-                            VStack(alignment: .leading, spacing: 4) {
-                                HStack(spacing: 10) {
-                                    Image("reactionLine")
-                                        .resizable()
-                                        .aspectRatio(contentMode: .fill)
-                                        .frame(width: 14, height: 12)
-
-                                    Text("내 답글")
-                                        .font(LDFont.medium05)
-                                        .foregroundColor(LDColor.color2)
-                                }
-
-                                Text(artistMessage.message)
-                                    .font(LDFont.medium03)
-                                    .foregroundColor(LDColor.color1)
-                                    .lineSpacing(4)
-                            }
-                        }
-                    }
-                    .padding(.top, 24)
-                }
-            }
-            .padding(.top, 12)
-            .padding(.bottom, 12)
-            .padding(.horizontal, 24)
-
-            if !isLast {
-                Rectangle()
-                    .fill(LDColor.gray8)
-                    .frame(height: 1)
-            }
-        }
-    }
-}
-
-// MARK: 작가 반응 확인뷰에서만 사용되는 이모지 팝업
-
-private struct EMojiPopupView: View {
-    let onSelect: (String) -> Void
-
-    var body: some View {
-        HStack(spacing: 12) {
-            ForEach(ReactionConstants.emojiAssets, id: \.self) { assetName in
-                Button(action: {
-                    onSelect(assetName)
-                }) {
-                    Image(assetName)
-                        .resizable()
-                        .scaledToFill()
-                        .frame(width: 36, height: 36)
-                        .padding(1)
-                }
-                .buttonStyle(PlainButtonStyle())
-            }
-            .shadow(color: .black.opacity(0.25), radius: 2, x: 0, y: 0)
-        }
-        .padding(.horizontal, 14)
-        .padding(.vertical, 8)
-        .background(.white)
-        .cornerRadius(40)
-        .shadow(color: .black.opacity(0.25), radius: 6, x: 1, y: 3)
-    }
-}
-
-// MARK: 작가 반응 확인뷰에서만 사용되는 반응 전송 팝업
-
-private struct MessagePopupView: View {
-    @Binding var showMessagePopup: Bool
-    @ObservedObject var viewModel: ArtworkReactionViewModel
-    let reaction: ReactionData?
-    @State private var messageText: String = ""
-    @FocusState private var isTextFieldFocused: Bool
-
-    var body: some View {
-        VStack(spacing: 12) {
-            Image("bubbleOff")
-                .resizable()
-                .aspectRatio(contentMode: .fill)
-                .frame(width: 27, height: 25)
-
-            Text("메시지에 반응해보세요")
-                .font(LDFont.heading04)
-                .foregroundStyle(LDColor.color1)
-                .lineSpacing(5)
-
-            TextField("10자 이내로 입력해주세요", text: $messageText)
-                .font(LDFont.medium03)
-                .foregroundStyle(LDColor.color1)
-                .padding(.horizontal, 12)
-                .padding(.vertical, 9)
-                .background(Color.white)
-                .cornerRadius(12)
-                .overlay(
-                    RoundedRectangle(cornerRadius: 12)
-                        .stroke(LDColor.color4, lineWidth: 1)
-                )
-                .focused($isTextFieldFocused)
-                .padding(.horizontal, 12)
-                .onChange(of: messageText) { newValue in
-                    if newValue.count > 10 {
-                        messageText = String(newValue.prefix(10))
-                    }
-                }
-
-            Spacer().frame(height: 10)
-
-            HStack(spacing: 9) {
-                PopupButton(
-                    title: "취소",
-                    foregroundColor: LDColor.color1,
-                    backgroundColor: LDColor.color4,
-                    action: {
-                        showMessagePopup = false
-                    }
-                )
-
-                PopupButton(
-                    title: "확인",
-                    foregroundColor: .white,
-                    backgroundColor: LDColor.color1,
-                    action: {
-                        guard let reaction = reaction else { return }
-                        viewModel.sendMessage(reaction.id, messageText) { success in
-                            if success {
-                                messageText = ""
-                                showMessagePopup = false
-                            }
-                        }
-                    }
-                )
-            }
-            .padding(.horizontal, 12)
-        }
-        .frame(width: 293)
-        .padding(.top, 28)
-        .padding(.bottom, 27)
-        .background(LDColor.color5)
-        .cornerRadius(14)
-    }
-}
-
-private struct PopupButton: View {
-    let title: String
-    let foregroundColor: Color
-    let backgroundColor: Color
-    let action: () -> Void
-
-    var body: some View {
-        Button(
-            action: action,
-            label: {
-                Text(title)
-                    .font(LDFont.heading06)
-                    .foregroundStyle(foregroundColor)
-                    .frame(maxWidth: .infinity, minHeight: 42)
-                    .background(backgroundColor)
-                    .cornerRadius(12)
-            }
-        )
-        .buttonStyle(PlainButtonStyle())
     }
 }
